Adds graphical interface
