+++ tests/test_spectra_calc_properties.py
@@ -8,7 +8,6 @@ from unittest import TestCase
 
 import numpy as np
 from astropy.constants import c
-from matplotlib import pyplot as plt
 from uncertainties.unumpy import uarray
 
 from phot_class import spectra
@@ -240,7 +239,7 @@ class FindPeakWavelength(TestCase):
 
         max_wavelength = max(self.wave)
         with self.assertRaises(ValueError):
-            spectra._calc_properties.find_peak_wavelength(
+            spectra.find_peak_wavelength(
                 wave=self.wave,
                 flux=self.flux,
                 lower_bound=max_wavelength + 10,
@@ -298,7 +297,7 @@ class FindFeatureBounds(TestCase):
             'upper_red': upper_peak_wavelength + 10
         }
 
-        feat_start, feat_end = spectra.find_feature_bounds(
+        feat_start, feat_end = spectra.guess_feature_bounds(
             wave, flux, feature_dict)
 
         self.assertEqual(
@@ -306,142 +305,6 @@ class FindFeatureBounds(TestCase):
 
         self.assertEqual(
             upper_peak_wavelength, feat_end, 'Incorrect max peak')
-
-
-class DrawMeasurement(TestCase):
-    """Tests for the ``_draw_measurement`` function"""
-
-    def runTest(self):
-        """Test the current axis is not empty after drawing"""
-
-        wave = np.arange(1000, 2000)
-        flux, error = SimulatedSpectrum.gaussian(wave, stddev=100)
-        spectra._calc_properties._draw_measurement(
-            wave, flux, flux, 'pW6', [2, 3, 4])
-
-        self.assertTrue(plt.gca(), 'Returned figure has no axes')
-
-
-class SampleFeatureProperties(TestCase):
-    """Tests for the ``sample_feature_properties`` function"""
-
-    @classmethod
-    def setUpClass(cls):
-        # Define test spectrum
-        cls.wave = np.arange(1000, 2000)
-        cls.observed_wave = np.mean(cls.wave)
-        cls.rest_wave = cls.observed_wave - 100
-        cls.flux, cls.error = SimulatedSpectrum.gaussian(
-            cls.wave, mean=cls.observed_wave, stddev=100)
-
-        # Select just the feature
-        cls.feat_start = cls.wave[100]
-        cls.feat_end = cls.wave[-100]
-        feat_wave = cls.wave[100:-100]
-        feat_flux = cls.flux[100:-100]
-
-        # Calculate feature properties
-        cls.area = spectra.feature_area(feat_wave, feat_flux)
-        continuum, norm_flux, cls.pequiv_width = \
-            spectra.feature_pew(feat_wave, feat_flux)
-
-        cls.velocity = spectra.feature_velocity(
-            cls.rest_wave, feat_wave, feat_flux)
-
-        cls.feat_name = 'test_CalcFeatureProperties'
-        line_properties = {
-            'restframe': cls.rest_wave,
-        }
-
-        spectra.line_locations[cls.feat_name] = line_properties
-
-    @classmethod
-    def tearDownClass(cls):
-        del spectra.line_locations[cls.feat_name]
-
-    def test_range_too_narrow(self):
-        """Test an error is raised if too narrow a feature is specified"""
-
-        args = dict(
-            feat_name=self.feat_name,
-            feat_start=self.wave[100],
-            feat_end=self.wave[109],
-            wave=self.wave,
-            flux=self.flux
-        )
-
-        self.assertRaisesRegex(
-            ValueError,
-            'Range too small.*',
-            spectra.sample_feature_properties,
-            **args
-        )
-
-    def assertNumberSamples(self, nstep, nsamp=None):
-        """Assert the correct number of samples were performed for a given
-         number of steps.
-
-         If ``nsamp`` is not specified use:
-             nsamp = ((2 * nstep) + 1) ** 2
-
-        Args:
-            nstep (int): Number of steps taken in each direction
-            nsamp (int): Number of expected samples
-        """
-
-        velocity, pequiv_width, area = spectra.sample_feature_properties(
-            self.feat_name, self.feat_start, self.feat_end, self.wave,
-            self.flux, nstep=nstep, debug=True)
-
-        msg = 'Wrong number of samples for n={}'
-        nsamp = ((2 * nstep) + 1) ** 2 if nsamp is None else nsamp
-        self.assertEqual(nsamp, len(velocity), msg.format(nstep))
-        self.assertEqual(nsamp, len(pequiv_width), msg.format(nstep))
-        self.assertEqual(nsamp, len(area), msg.format(nstep))
-
-    def test_number_of_samples(self):
-        """Test the correct number of samples are performed"""
-
-        for n in [0, 5]:
-            self.assertNumberSamples(n)
-
-        self.assertNumberSamples(0, 1)
-        self.assertNumberSamples(-1, 0)
-
-    def test_return_order(self):
-        """Test values are returned in the correct order"""
-
-        velocity, pequiv_width, area = spectra.sample_feature_properties(
-            self.feat_name, self.feat_start, self.feat_end, self.wave,
-            self.flux, nstep=0)
-
-        msg = '{} values do not match.'
-        # Todo: Add back in the velocity calculation
-        # self.assertAlmostEqual(
-        #     self.velocity, velocity[0], msg=msg.format('Velocity'))
-
-        self.assertAlmostEqual(
-            self.pequiv_width, pequiv_width[0], msg=msg.format('pEW'))
-
-        self.assertAlmostEqual(self.area, area[0], msg=msg.format('Area'))
-
-    def test_uarray_support(self):
-        """Test the function supports input arrays with ufloat objects"""
-
-        uflux = uarray(self.flux, self.error)
-        velocity, pequiv_width, area = spectra.sample_feature_properties(
-            self.feat_name, self.feat_start, self.feat_end, self.wave, uflux,
-            nstep=0)
-
-        msg = '{} values do not match.'
-        # Todo: Add back in the velocity calculation
-        # self.assertAlmostEqual(
-        #     self.velocity, velocity[0], msg=msg.format('Velocity'))
-
-        self.assertAlmostEqual(
-            self.pequiv_width, pequiv_width[0], msg=msg.format('pEW'))
-
-        self.assertAlmostEqual(self.area, area[0], msg=msg.format('Area'))
 
 
 class CorrectSpectrum(TestCase):
@@ -466,7 +329,7 @@ class CorrectSpectrum(TestCase):
         rv = 3.1
 
         shifted_wave = self.test_wave * (1 + z)
-        rested_wave, flux = spectra._calc_properties.correct_extinction(
+        rested_wave, flux = spectra.correct_extinction(
             shifted_wave, self.test_flux, ra, dec, z=z, rv=rv
         )
         self.assertListEqual(
@@ -493,7 +356,7 @@ class CorrectSpectrum(TestCase):
         ext = extinction.fitzpatrick99(self.test_wave, a_v=rv * mwebv)
         extincted_flux = extinction.apply(ext, self.test_flux)
 
-        wave, flux = spectra._calc_properties.correct_extinction(
+        wave, flux = spectra.correct_extinction(
             self.test_wave, extincted_flux, ra, dec, z, rv=rv)
 
         is_close = np.isclose(self.test_flux, flux).all()
@@ -569,8 +432,8 @@ class TabulateSpectrumProperties(TestCase):
             'upper_red': cls.wave[-100]
         }
 
-        cls.old_lines = spectra._calc_properties.line_locations
-        spectra._calc_properties.line_locations = \
+        cls.old_lines = spectra.calc_properties.line_locations
+        spectra.calc_properties.line_locations = \
             {'test_TabulateSpectrumProperties': line_properties}
 
     @classmethod
